feat: implement sys_open, sys_close, sys_read, sys_write, and sys_lseek functions for file operations
