+++ kernel/src/task/syscall.rs
@@ -285,182 +285,3 @@ pub fn sys_getppid(trapframe: &mut Trapframe) -> usize {
     trapframe.epc += 4;
     task.get_parent_id().unwrap_or(task.get_id()) as usize
 }
-
-pub fn sys_open(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let path_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(0)).unwrap() as *const u8;
-    let _flags = trapframe.get_arg(1) as i32;
-    let _mode = trapframe.get_arg(2) as i32;
-
-    // Increment PC to avoid infinite loop if open fails
-    trapframe.epc += 4;
-
-    // Parse path as a null-terminated C string
-    let mut path_bytes = Vec::new();
-    let mut i = 0;
-    unsafe {
-        loop {
-            let byte = *path_ptr.add(i);
-            if byte == 0 {
-                break;
-            }
-            path_bytes.push(byte);
-            i += 1;
-
-            if i > MAX_PATH_LENGTH {
-                return usize::MAX; // Path too long
-            }
-        }
-    }
-
-    // Convert path bytes to string
-    let path_str = match str::from_utf8(&path_bytes) {
-        Ok(s) => s,
-        Err(_) => return usize::MAX, // Invalid UTF-8
-    };
-
-    // Try to open the file
-    let file = File::open(path_str.to_string());
-    match file {
-        Ok(file) => {
-            // Register the file with the task
-            let fd = task.add_file(file);
-            if fd.is_err() {
-                return usize::MAX; // File descriptor error
-            }
-            fd.unwrap() as usize
-        }
-        Err(_) => usize::MAX, // File open error
-    }
-}
-
-pub fn sys_close(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as usize;
-    trapframe.epc += 4;
-    if task.remove_file(fd).is_ok() {
-        0
-    } else {
-        usize::MAX // -1
-    }
-}
-
-pub fn sys_read(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as usize;
-    let buf_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *mut u8;
-    let count = trapframe.get_arg(2) as usize;
-
-    // Increment PC to avoid infinite loop if read fails
-    trapframe.epc += 4;
-
-    let file = task.get_mut_file(fd);
-    if file.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let file = file.unwrap();
-
-    let buffer = unsafe { core::slice::from_raw_parts_mut(buf_ptr, count) };
-    
-    match file.read(buffer) {
-        Ok(n) => {
-            n
-        }
-        Err(_) => {
-            return usize::MAX; // Read error
-        }
-    }
-}
-
-pub fn sys_write(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as usize;
-    let buf_ptr = task.vm_manager.translate_vaddr(trapframe.get_arg(1)).unwrap() as *const u8;
-    let count = trapframe.get_arg(2) as usize;
-
-    // Increment PC to avoid infinite loop if write fails
-    trapframe.epc += 4;
-
-    let file = task.get_mut_file(fd);
-    if file.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let file = file.unwrap();
-
-    let buffer = unsafe { core::slice::from_raw_parts(buf_ptr, count) };
-    
-    match file.write(buffer) {
-        Ok(n) => {
-            n
-        }
-        Err(_) => {
-            return usize::MAX; // Write error
-        }
-    }
-}
-
-pub fn sys_lseek(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as usize;
-    let offset = trapframe.get_arg(1) as i32;
-    let whence = trapframe.get_arg(2) as i32;
-
-    // Increment PC to avoid infinite loop if lseek fails
-    trapframe.epc += 4;
-
-    let file = task.get_mut_file(fd);
-    if file.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let file = file.unwrap();
-    let whence  = match whence {
-        0 => SeekFrom::Start(offset as u64),
-        1 => SeekFrom::Current(offset as i64),
-        2 => SeekFrom::End(offset as i64),
-        _ => return usize::MAX, // Invalid whence
-    };
-
-    match file.seek(whence) {
-        Ok(pos) => {
-            pos as usize
-        }
-        Err(_) => {
-            return usize::MAX; // Lseek error
-        }
-    }
-}
-
-pub fn sys_lseek64(trapframe: &mut Trapframe) -> usize {
-    let task = mytask().unwrap();
-    let fd = trapframe.get_arg(0) as usize;
-    let offset = trapframe.get_arg(1) as i64;
-    let whence = trapframe.get_arg(2) as i32;
-
-    // Increment PC to avoid infinite loop if lseek fails
-    trapframe.epc += 4;
-
-    let file = task.get_mut_file(fd);
-    if file.is_none() {
-        return usize::MAX; // Invalid file descriptor
-    }
-
-    let file = file.unwrap();
-    let whence  = match whence {
-        0 => SeekFrom::Start(offset as u64),
-        1 => SeekFrom::Current(offset),
-        2 => SeekFrom::End(offset),
-        _ => return usize::MAX, // Invalid whence
-    };
-
-    match file.seek(whence) {
-        Ok(pos) => {
-            pos as usize
-        }
-        Err(_) => {
-            return usize::MAX; // Lseek error
-        }
-    }
-}
